requirements.txt
gradio>=4.0.0
torch>=2.0.0
transformers>=4.30.0
python-ollama>=0.1.0
Pillow>=10.0.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
